requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-storage
firebase-admin 
jinja2
